requirements.txt
grpcio
grpcio-tools
grpcio-health-checking

protobuf<7.0.0,>=6.30.0
fastapi
uvicorn
